Adds HTMLUrlFinder object to find and extract urls on a visited page
These 'allowed' urls are feed back into the internal url/job queue managed by crwlr
Allowed urls are urls that point to the same baseUrl.
For example, given a domain example.com:
- example.com/foo is allowed
- sub.example.com/foo is not allowed
- otherexample.com/bar is not allowed and will not be visited

The HtmlUrlFinder is created as a plugin and subscribes to 'crwlr.visitor.post-visit' event published when a page is visited.